blob minimum area now dependent on current sensor image dimensions.

+++ openframeworks/wiretouch_monitor/src/blob-tracker.h
@@ -52,8 +52,8 @@ public:
     vector<ofxStoredBlobVO>& currentBlobs();
     
 protected:
-    int                     blobsMinArea;
-    int                     blobsMaxArea;
+    float                   blobsMinArea;
+    float                   blobsMaxArea;
     int                     blobsNumMax;
     int                     adaptiveThreshRange;
